Fix Image.getColor() for scaled images
Although the behaviour is questionable, since the underlying texture can be
bigger than the actual image, and it's not stretched when the image is scaled.
Instead, it repeats, but only on the texture dimensions (not the image ones)
which are always the next power of two of the size of the image.

To sum up, I don't think this method makes a lot of sense, but at least it
doesn't crash anymore.

Original issue: https://bitbucket.org/kevglass/slick/issue/42/getcolor-crashes-for-a-flipped-image