Merge development branch from upstream
Note while merging Image.java:
This changes Image.getScaledCopy to behave as in b235068e and produces a
resampled image upon drawing. Image.getColor still does not account for scaled
images but no longer throws ArrayOutOfBoundsExcepeption when getting new
coordinates, instead it returns the unscaled color and if coordinates are
larger than the original texture, the texture wraps.

This merely fixes the exception that arrises and does not the correct
Image.getColor to be logically correct. The proper fix to Image.getColor would
be to return the color as rendered by the Image draw call, regardless of the
underlying implementation.

Conflicts:
	slick2d-core/src/main/java/org/newdawn/slick/Image.java
	slick2d-core/src/main/java/org/newdawn/slick/geom/Rectangle.java
	slick2d-core/src/test/java/org/newdawn/slick/tests/UnicodeFontTest.java